+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/model/MeetingModel.kt
@@ -35,6 +35,7 @@ class MeetingModel : ViewModel() {
     var isCameraOn = false
     var isDeviceListDialogOn = false
     var isAdditionalOptionsDialogOn = false
+    var isSharingContent = false
     var lastReceivedMessageTimestamp = 0L
     var tabIndex = 0
     var isUsingCameraCaptureSource = true